refactor(ui): removed old ui code
