+++ pkg/ui/props.go
@@ -1,0 +1,6 @@
+package ui
+
+type TextDisplayProperties struct {
+	Text   string `json:"text"`
+	Format string `json:"format"`
+}
